Adicionado arquivo Makefile, corrigido função divi

+++ src/divi.c
@@ -1,6 +1,5 @@
-double divi(int a, int b){
+double divi(double a, double b){
     double x;
-    
     x = a / b;
     return x;
 }
