+++ src/functions.h
@@ -3,6 +3,6 @@
 int add (int a, int b);
 int sub(int a, int b);
 int mul(int a, int b);
-double divi(int a, int b);
+double divi(double a, double b);
 double raiz(double a);
 int pot(int a, int b);
